move ws files up one level
