+++ sdk/servicebus/azure-servicebus/azure/servicebus/_pyamqp/_constants.py
