+++ sdk/servicebus/azure-servicebus/azure/servicebus/_pyamqp/_headers.py
