+++ sdk/servicebus/azure-servicebus/azure/servicebus/_pyamqp/_utils.py
