+++ sdk/servicebus/azure-servicebus/azure/servicebus/_pyamqp/websockets.py
